added getting playerID out of context token
edited validation for claim rights in service layer for gameRoom/Vote/Player/Round services

+++ ScrumPoker.Business/GameRoomService.cs
@@ -1,5 +1,6 @@
 using ScrumPoker.Business.Interfaces.Interfaces;
 using ScrumPoker.Business.Models.Models;
+using ScrumPoker.Common.ConflictExceptions;
 using ScrumPoker.DataAccess.Interfaces;
 
 namespace ScrumPoker.Business;
@@ -8,10 +9,12 @@ namespace ScrumPoker.Business;
 public class GameRoomService : IGameRoomService
 {
     private readonly IGameRoomRepository _gameRoomRepository;
+    private readonly IUserManager _userManager;
 
-    public GameRoomService(IGameRoomRepository gameRoomRepository)
+    public GameRoomService(IGameRoomRepository gameRoomRepository, IUserManager userManager)
     {
         _gameRoomRepository = gameRoomRepository;
+        _userManager = userManager;
     }
 
     public async Task<List<GameRoom>> GetAll()
@@ -26,9 +29,9 @@ public class GameRoomService : IGameRoomService
 
     public async Task<GameRoom> Create(GameRoom gameRoomRequest, int requestId)
     {
-        var gameRoom = gameRoomRequest;
-        gameRoom.MasterId = requestId;
-        var gameRooms = await _gameRoomRepository.Create(gameRoom);
+        var masterId = _userManager.GetUserId();
+        gameRoomRequest.MasterId = masterId;
+        var gameRooms = await _gameRoomRepository.Create(gameRoomRequest);
         await _gameRoomRepository.AddPlayerToRoom(gameRooms.Id, gameRooms.MasterId);
         var gameRoomResponse = await _gameRoomRepository.GetById(gameRooms.Id);
 
@@ -37,6 +40,13 @@ public class GameRoomService : IGameRoomService
 
     public async Task<GameRoom> Update(GameRoom gameRoomRequest)
     {
+        var gameRoomDto = await GetById(gameRoomRequest.Id);
+        var masterId = _userManager.GetUserId();
+        if (gameRoomDto.MasterId != masterId)
+        {
+            throw new HasNoClaimException($"User has not rights to Update game room (ID {gameRoomRequest.Id})");
+        }
+        
         return await _gameRoomRepository.Update(gameRoomRequest);
     }
 
@@ -52,11 +62,25 @@ public class GameRoomService : IGameRoomService
 
     public async Task AddPlayer(int gameRoomId, int playerId)
     {
+        var gameRoomDto = await GetById(gameRoomId);
+        var masterId = _userManager.GetUserId();
+        if (gameRoomDto.MasterId != masterId)
+        {
+            throw new HasNoClaimException($"User has not rights to Update game room (ID {gameRoomId})");
+        }
+        
         await _gameRoomRepository.AddPlayerToRoom(gameRoomId, playerId);
     }
 
     public async Task RemovePlayer(int gameRoomId, int playerId)
     {
+        var gameRoomDto = await GetById(gameRoomId);
+        var masterId = _userManager.GetUserId();
+        if (gameRoomDto.MasterId != masterId)
+        {
+            throw new HasNoClaimException($"User has not rights to Update game room (ID {gameRoomId})");
+        }
+
         await _gameRoomRepository.RemovePlayerById(gameRoomId, playerId);
     }
 }
